+++ sample-apps/pathology/lib/infers/__init__.py
@@ -8,10 +8,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .handlers import TensorBoardImageHandler
-from .infers.deepedit import InferDeepedit
-from .infers.segmentation import InferSegmentation
-from .trainers.deepedit import TrainDeepEdit
-from .trainers.segmentation import TrainSegmentation
-from .utils import split_pannuke_dataset
