+++ sample-apps/pathology/lib/infers/segmentation.py
@@ -8,12 +8,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 import logging
-from typing import Callable, Sequence
+from typing import Any, Callable, Dict, Sequence
 
 import numpy as np
-from monai.apps.deepgrow.transforms import AddGuidanceSignald
+from lib.transforms import FilterImaged, FindContoursd, LoadImagePatchd, PostFilterLabeld
 from monai.transforms import (
     Activationsd,
     AsChannelFirstd,
@@ -24,15 +23,13 @@ from monai.transforms import (
     ToNumpyd,
 )
 
-from monailabel.interfaces.tasks.infer import InferType
-
-from .infer import InferSegmentation
-from .transforms import AddClickGuidanced, FilterImaged, FindContoursd, LoadImagePatchd, PostFilterLabeld
+from monailabel.interfaces.tasks.infer import InferTask, InferType
+from monailabel.transform.writer import PolygonWriter
 
 logger = logging.getLogger(__name__)
 
 
-class InferDeepedit(InferSegmentation):
+class InferSegmentation(InferTask):
     """
     This provides Inference Engine for pre-trained segmentation (UNet) model over MSD Dataset.
     """
@@ -42,10 +39,10 @@ class InferDeepedit(InferSegmentation):
         path,
         network=None,
         roi_size=(256, 256),
-        type=InferType.DEEPEDIT,
+        type=InferType.SEGMENTATION,
         labels=None,
         dimension=2,
-        description="A pre-trained interaction/deepedit model for Nuclei (Pathology)",
+        description="A pre-trained semantic segmentation model for Pathology",
     ):
         super().__init__(
             path=path,
@@ -57,23 +54,33 @@ class InferDeepedit(InferSegmentation):
             description=description,
         )
 
-    def pre_transforms(self, data=None):
+    def info(self) -> Dict[str, Any]:
+        d = super().info()
+        d["pathology"] = True
+        return d
+
+    def pre_transforms(self, data=None) -> Sequence[Callable]:
         return [
             LoadImagePatchd(keys="image", conversion="RGB", dtype=np.uint8),
             FilterImaged(keys="image"),
             AsChannelFirstd(keys="image"),
             ScaleIntensityRangeD(keys="image", a_min=0.0, a_max=255.0, b_min=-1.0, b_max=1.0),
-            AddClickGuidanced(image="image", guidance="guidance"),
-            AddGuidanceSignald(image="image", guidance="guidance", number_intensity_ch=3),
             EnsureTyped(keys="image"),
         ]
 
     def post_transforms(self, data=None) -> Sequence[Callable]:
         return [
-            Activationsd(keys="pred", sigmoid=True),
-            AsDiscreted(keys="pred", threshold=0.5),
-            SqueezeDimd(keys="pred"),
+            Activationsd(keys="pred", softmax=True),
+            AsDiscreted(
+                keys="pred",
+                argmax=True,
+            ),
+            SqueezeDimd(keys="pred", dim=0),
             ToNumpyd(keys=("image", "pred")),
             PostFilterLabeld(keys="pred", image="image"),
             FindContoursd(keys="pred", labels=self.labels),
         ]
+
+    def writer(self, data, extension=None, dtype=None):
+        writer = PolygonWriter(label=self.output_label_key, json=self.output_json_key)
+        return writer(data)
